make form state modular
